+++ ord-svr-feign/src/main/java/rebue/ord/svr/feign/OrdTaskSvc.java
@@ -11,11 +11,13 @@ import rebue.ord.mo.OrdTaskMo;
 import rebue.sbs.feign.FeignConfig;
 
 /**
- * 创建时间：2018年5月21日 下午3:21:05 项目名称：ord-svr-feign
+ * 创建时间：2018年5月21日 下午3:21:05 
+ * 项目名称：ord-svr-feign
  * 
  * @author daniel
  * @version 1.0
- * @since JDK 1.8 文件名称：OrdTaskSvc.java 类说明： 订单任务内部接口
+ * @since JDK 1.8 文件名称：OrdTaskSvc.java 
+ * 类说明： 订单任务内部接口
  */
 @FeignClient(name = "ord-svr", configuration = FeignConfig.class)
 public interface OrdTaskSvc {
@@ -27,8 +29,8 @@ public interface OrdTaskSvc {
 	 * @return
 	 * @date 2018年5月21日 下午3:24:56
 	 */
-	@GetMapping(value = "/ord/task")
-	List<OrdTaskMo> list(@RequestParam("executeState") byte executeState, @RequestParam("taskType") byte taskType);
+	@GetMapping(value = "/ord/task") 
+	List<OrdTaskMo> list(@RequestParam("executeState") int executeState, @RequestParam("taskType") int taskType);
 
 	/**
 	 * 执行订单签收任务 Title: executeSignInOrderTask Description:
